Rename userIdentificationService
