+++ src/main/java/uk/gov/hmcts/reform/draftstore/service/AuthService.java
@@ -13,7 +13,7 @@ For tactical reasons, no IDAM stubs/not deployed anywhere, Divorce BA's decided 
 Initially the implementation did indeed send a request to IDAM's /details endpoint to get the documenter's user id
 @see http://git.reform/divorce/draft-document-store/commit/91839f84e54e42f5b70f28c68a150ad60ef86b9e
  */
-public class UserIdentificationService {
+public class AuthService {
 
     public static final String SERVICE_HEADER = "ServiceAuthorization";
     public static final String AUTH_TYPE = "hmcts-id ";
